fix(api): cast rlim_cur to u64 for 32-bit Linux compat

+++ crates/meow-api/src/routes.rs
@@ -2032,7 +2032,7 @@ async fn read_os_memory_limit_linux() -> u64 {
             rlim_max: 0,
         };
         if libc::getrlimit(libc::RLIMIT_AS, &mut rl) == 0 && rl.rlim_cur != libc::RLIM_INFINITY {
-            return rl.rlim_cur;
+            return rl.rlim_cur as u64;
         }
     }
     0
